Refactor login method to serices
